+++ src/models/mistral_hf.py
@@ -42,33 +42,33 @@ class MistralAttentionHf(AttentionForCausalLM):
             bias=False,
             gather_output=False,
             init_method=lambda x: x,
-        )
+        ).type(self.args.dtype)
         self.k_proj = ColumnParallelLinear(
             self.args.hidden_size,
             self.args.num_key_value_heads * self.head_dim,
             bias=False,
             gather_output=False,
             init_method=lambda x: x,
-        )
+        ).type(self.args.dtype)
         self.v_proj = ColumnParallelLinear(
             self.args.hidden_size,
             self.args.num_key_value_heads * self.head_dim,
             bias=False,
             gather_output=False,
             init_method=lambda x: x,
-        )
+        ).type(self.args.dtype)
         self.o_proj = RowParallelLinear(
             self.args.num_attention_heads * self.head_dim,
             self.args.hidden_size,
             bias=False,
             input_is_parallel=True,
             init_method=lambda x: x,
-        )
+        ).type(self.args.dtype)
         self.rotary_emb = RotaryEmbedding(
             self.head_dim,
             max_position_embeddings=self.args.max_position_embeddings,
             base=self.args.rope_theta
-        )
+        ).type(self.args.dtype)
 
     def forward(
             self,
@@ -116,21 +116,21 @@ class MistralFeedForwardHf(nn.Module):
             bias=False,
             gather_output=False,
             init_method=lambda x: x
-        )
+        ).type(self.args.dtype)
         self.down_proj = RowParallelLinear(
             self.args.intermediate_size,
             self.args.hidden_size,
             bias=False,
             input_is_parallel=True,
             init_method=lambda x: x
-        )
+        ).type(self.args.dtype)
         self.up_proj = ColumnParallelLinear(
             self.args.hidden_size,
             self.args.intermediate_size,
             bias=False,
             gather_output=False,
             init_method=lambda x: x
-        )
+        ).type(self.args.dtype)
 
     def forward(self, x) -> torch.Tensor:
         return self.down_proj(F.silu(self.gate_proj(x)) * self.up_proj(x))
@@ -150,8 +150,8 @@ class MistralTransformerBlockHf(nn.Module):
     def init_weights(self):
         self.self_attn.init_weights()
         self.mlp.init_weights()
-        self.input_layernorm = RMSNorm(self.args.hidden_size, eps=self.args.rms_norm_eps)
-        self.post_attention_layernorm = RMSNorm(self.args.hidden_size, eps=self.args.rms_norm_eps)
+        self.input_layernorm = RMSNorm(self.args.hidden_size, eps=self.args.rms_norm_eps).type(self.args.dtype)
+        self.post_attention_layernorm = RMSNorm(self.args.hidden_size, eps=self.args.rms_norm_eps).type(self.args.dtype)
 
     def forward(
             self,
@@ -181,10 +181,10 @@ class MistralModelHf(nn.Module):
     def init_weights(self):
         self.embed_tokens = ParallelEmbedding(
             self.args.vocab_size, self.args.hidden_size, init_method=lambda x: x
-        )
+        ).type(self.args.dtype)
         for layer in self.layers:
             layer.init_weights()
-        self.norm = RMSNorm(self.args.hidden_size, eps=self.args.rms_norm_eps)
+        self.norm = RMSNorm(self.args.hidden_size, eps=self.args.rms_norm_eps).type(self.args.dtype)
 
     def forward(
             self,
@@ -217,7 +217,7 @@ class MistralHf(ParallelModelForCausalLM):
         self.model.init_weights()
         self.lm_head = ColumnParallelLinear(
             self.args.hidden_size, self.args.vocab_size, bias=False, init_method=lambda x: x
-        )
+        ).type(self.args.dtype)
 
     def forward(self, tokens: torch.Tensor, start_pos=0, use_cache=False):
         h = self.model.forward(tokens, start_pos, use_cache)
